Make save-state keybindings configurable via keybindings.toml
The save/load/slot keys added for save states are now part of the `[app]`
section of keybindings.toml instead of being hardcoded:

    [app]
    fps_toggle = "KeyF"
    save_state = "F5"
    load_state = "F9"
    slots = ["Digit0", ..., "Digit9"]   # exactly 10 keys

- input.rs: RawAppConfig gains save_state, load_state, and a fixed-size
  slots: [KeyCode; NUM_SLOTS] array; every [app] field now has an independent
  serde default, so an older keybindings.toml missing any of them (or missing
  [app] entirely) still parses. KeyMap exposes is_save_state / is_load_state /
  slot_for_key.
- main.rs: the keyboard handler consults KeyMap for the slot/save/load keys;
  the hardcoded digit_slot / F5 / F9 handling is removed.
- app.rs: NUM_SLOTS is now the shared usize slot count (used to size the
  keybinding array).
- assets/keybindings.toml: ship the new [app] keys.

The number keys 0-9 / F5 / F9 remain the defaults, so behavior is unchanged
out of the box. build.rs never overwrites an existing keybindings.toml, and
the per-field defaults keep those older files working.

Tests (input.rs): defaults expose F5/F9 + Digit0-9; a full [app] override
(incl. a Numpad slots array) takes effect; each [app] field defaults
independently; a slots array of the wrong length is a parse error. The
toml/serde array+default mechanics were also cross-checked standalone.

+++ src/app.rs
@@ -32,8 +32,9 @@ pub struct App {
     active_slot: u8,
 }
 
-/// Number of numbered save-state slots (selectable with keys 0-9).
-pub const NUM_SLOTS: u8 = 10;
+/// Number of numbered save-state slots (selectable with the configured slot
+/// keys, by default the number-row keys 0-9).
+pub const NUM_SLOTS: usize = 10;
 
 /// File path for save-state `slot` given the loaded ROM path. Slot 0 uses
 /// `<rom>.state` (the original single-slot name, kept for backward
@@ -249,7 +250,7 @@ impl App {
     /// Select the active save-state slot (used by the number keys). Out-of-range
     /// values are ignored. Updates the window title to reflect the new slot.
     pub fn select_slot(&mut self, slot: u8) {
-        if slot >= NUM_SLOTS {
+        if slot as usize >= NUM_SLOTS {
             return;
         }
         self.active_slot = slot;
@@ -347,7 +348,7 @@ mod tests {
             slot_state_path(Some(&rom), 0),
             Some(PathBuf::from("/games/mario.state"))
         );
-        for slot in 1..NUM_SLOTS {
+        for slot in 1..NUM_SLOTS as u8 {
             assert_eq!(
                 slot_state_path(Some(&rom), slot),
                 Some(PathBuf::from(format!("/games/mario.state{slot}"))),
